backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
google-cloud-storage==2.8.0
python-dotenv==0.20.0

